adds accordion with quadarnt data in radarview
